added alter table query to add auto increment in block user table and done proper squences of sql file in db.changlog-master file

+++ src/main/resources/db/changelog/tables/create_table-users.sql
@@ -10,7 +10,7 @@ CREATE TABLE users
     email VARCHAR(255),
     password VARCHAR(255),
     address VARCHAR(255),
-    phone_number INT,
+    phone_number BIGINT,
     birth_date date,
     created_at date,
     updated_at date,
